requirements.txt
slack_sdk>=3.0
python-dotenv~=0.15.0